ACPI: APEI: Don't warn if ACPI is disabled
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Jon Hunter <jonathanh@nvidia.com>
commit b7a732a73ac54dbccf14d90779d7e6b70ea0d653

If ACPI is not enabled but support for ACPI and APEI is enabled in the
kernel, then the following warning is seen on boot ...

 WARNING KERN EINJ: ACPI disabled.

For ARM64 platforms, the 'acpi_disabled' variable is true by default
and hence, the above is often seen on ARM64. Given that it can be
normal for ACPI to be disabled, make this an informational print rather
that a warning.

	Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
	Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit b7a732a73ac54dbccf14d90779d7e6b70ea0d653)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>