netfilter: nf_tables: relocate header content to consumer
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Paul Gortmaker <paul.gortmaker@windriver.com>
commit f1f86d09ca7e35fb161a47bc54ec9cb2f4fe42d8

The nf_tables.h header is used in a lot of files, but it turns out
that there is only one actual user of nft_expr_clone().

Hence we relocate that function to be with the one consumer of it
and avoid having to process it with CPP for all the other files.

This will also enable a reduction in the other headers that the
nf_tables.h itself has to include just to be stand-alone, hence
a pending further significant reduction in the CPP content that
needs to get processed for each netfilter file.

Note that the explicit "inline" has been dropped as part of this
relocation.  In similar changes to this, I believe Dave has asked
this be done, so we free up gcc to make the choice of whether to
inline or not.

	Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
	Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>
(cherry picked from commit f1f86d09ca7e35fb161a47bc54ec9cb2f4fe42d8)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>