hugetlb_cgroup: add accounting for shared mappings
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Mina Almasry <almasrymina@google.com>
commit 075a61d07a8eca2fe980acd94105ed5d6429c55d

For shared mappings, the pointer to the hugetlb_cgroup to uncharge lives
in the resv_map entries, in file_region->reservation_counter.

After a call to region_chg, we charge the approprate hugetlb_cgroup, and
if successful, we pass on the hugetlb_cgroup info to a follow up
region_add call.  When a file_region entry is added to the resv_map via
region_add, we put the pointer to that cgroup in
file_region->reservation_counter.  If charging doesn't succeed, we report
the error to the caller, so that the kernel fails the reservation.

On region_del, which is when the hugetlb memory is unreserved, we also
uncharge the file_region->reservation_counter.

[akpm@linux-foundation.org: forward declare struct file_region]
	Signed-off-by: Mina Almasry <almasrymina@google.com>
	Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
	Reviewed-by: Mike Kravetz <mike.kravetz@oracle.com>
	Cc: David Rientjes <rientjes@google.com>
	Cc: Greg Thelen <gthelen@google.com>
	Cc: Mike Kravetz <mike.kravetz@oracle.com>
	Cc: Sandipan Das <sandipan@linux.ibm.com>
	Cc: Shakeel Butt <shakeelb@google.com>
	Cc: Shuah Khan <shuah@kernel.org>
Link: http://lkml.kernel.org/r/20200211213128.73302-5-almasrymina@google.com
	Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
(cherry picked from commit 075a61d07a8eca2fe980acd94105ed5d6429c55d)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>