bnxt_en: Get the RX packet timestamp
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Pavan Chebbi <pavan.chebbi@broadcom.com>
commit 7f5515d19cd7aa02a866fd86622a022f12e06f0f

If the RX packet is timestamped by the hardware, the RX completion
record will contain the lower 32-bit of the timestamp.  This needs
to be combined with the upper 16-bit of the periodic timestamp that
we get from the timer.  The previous snapshot in ptp->old_timer is
used to make sure that the snapshot is not ahead of the RX timestamp
and we adjust for wrap-around if needed.

v2: Make ptp->old_time read access safe on 32-bit CPUs.

	Reviewed-by: Edwin Peer <edwin.peer@broadcom.com>
	Signed-off-by: Pavan Chebbi <pavan.chebbi@broadcom.com>
	Signed-off-by: Michael Chan <michael.chan@broadcom.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 7f5515d19cd7aa02a866fd86622a022f12e06f0f)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>