bnxt: make sure xmit_more + errors does not miss doorbells
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Jakub Kicinski <kuba@kernel.org>
commit e8d8c5d80f5e9d4586c68061b62c642752289095

skbs are freed on error and not put on the ring. We may, however,
be in a situation where we're freeing the last skb of a batch,
and there is a doorbell ring pending because of xmit_more() being
true earlier. Make sure we ring the door bell in such situations.

Since errors are rare don't pay attention to xmit_more() and just
always flush the pending frames.

The busy case should be safe to be left alone because it can
only happen if start_xmit races with completions and they
both enable the queue. In that case the kick can't be pending.

Noticed while reading the code.

Fixes: 4d172f21cefe ("bnxt_en: Implement xmit_more.")
	Reviewed-by: Michael Chan <michael.chan@broadcom.com>
	Reviewed-by: Edwin Peer <edwin.peer@broadcom.com>
	Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit e8d8c5d80f5e9d4586c68061b62c642752289095)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>