mm: Allow non-direct-map arguments to free_reserved_area()
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Dave Hansen <dave.hansen@linux.intel.com>
commit 0d83432811f26871295a9bc24d3c387924da6071

free_reserved_area() takes pointers as arguments to show which addresses
should be freed.  However, it does this in a somewhat ambiguous way.  If it
gets a kernel direct map address, it always works.  However, if it gets an
address that is part of the kernel image alias mapping, it can fail.

It fails if all of the following happen:
 * The specified address is part of the kernel image alias
 * Poisoning is requested (forcing a memset())
 * The address is in a read-only portion of the kernel image

The memset() fails on the read-only mapping, of course.
free_reserved_area() *is* called both on the direct map and on kernel image
alias addresses.  We've just lucked out thus far that the kernel image
alias areas it gets used on are read-write.  I'm fairly sure this has been
just a happy accident.

It is quite easy to make free_reserved_area() work for all cases: just
convert the address to a direct map address before doing the memset(), and
do this unconditionally.  There is little chance of a regression here
because we previously did a virt_to_page() on the address for the memset,
so we know these are not highmem pages for which virt_to_page() would fail.

	Signed-off-by: Dave Hansen <dave.hansen@linux.intel.com>
	Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
	Cc: keescook@google.com
	Cc: aarcange@redhat.com
	Cc: jgross@suse.com
	Cc: jpoimboe@redhat.com
	Cc: gregkh@linuxfoundation.org
	Cc: peterz@infradead.org
	Cc: hughd@google.com
	Cc: torvalds@linux-foundation.org
	Cc: bp@alien8.de
	Cc: luto@kernel.org
	Cc: ak@linux.intel.com
	Cc: Kees Cook <keescook@google.com>
	Cc: Andrea Arcangeli <aarcange@redhat.com>
	Cc: Juergen Gross <jgross@suse.com>
	Cc: Josh Poimboeuf <jpoimboe@redhat.com>
	Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
	Cc: Peter Zijlstra <peterz@infradead.org>
	Cc: Hugh Dickins <hughd@google.com>
	Cc: Linus Torvalds <torvalds@linux-foundation.org>
	Cc: Borislav Petkov <bp@alien8.de>
	Cc: Andy Lutomirski <luto@kernel.org>
	Cc: Andi Kleen <ak@linux.intel.com>
Link: https://lkml.kernel.org/r/20180802225826.1287AE3E@viggo.jf.intel.com

(cherry picked from commit 0d83432811f26871295a9bc24d3c387924da6071)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>