swiotlb: Validate bounce size in the sync/unmap path
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Martin Radev <martin.b.radev@gmail.com>
commit daf9514fd5eb098d7d6f3a1247cb8cc48fc94155

The size of the buffer being bounced is not checked if it happens
to be larger than the size of the mapped buffer. Because the size
can be controlled by a device, as it's the case with virtio devices,
this can lead to memory corruption.

This patch saves the remaining buffer memory for each slab and uses
that information for validation in the sync/unmap paths before
swiotlb_bounce is called.

Validating this argument is important under the threat models of
AMD SEV-SNP and Intel TDX, where the HV is considered untrusted.

	Signed-off-by: Martin Radev <martin.b.radev@gmail.com>
	Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
(cherry picked from commit daf9514fd5eb098d7d6f3a1247cb8cc48fc94155)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>