bus: mhi: core: Clear context for stopped channels from remove()
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Bhaumik Bhatt <bbhatt@codeaurora.org>
commit 4e44ae3d6d9c2c2a6d9356dd279c925532d5cd8c

If a channel was explicitly stopped but not reset and a driver
remove is issued, clean up the channel context such that it is
reflected on the device. This move is useful if a client driver
module is unloaded or a device crash occurs with the host having
placed the channel in a stopped state.

	Signed-off-by: Bhaumik Bhatt <bbhatt@codeaurora.org>
	Reviewed-by: Hemant Kumar <hemantk@codeaurora.org>
	Reviewed-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
Link: https://lore.kernel.org/r/1617311778-1254-3-git-send-email-bbhatt@codeaurora.org
	Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
(cherry picked from commit 4e44ae3d6d9c2c2a6d9356dd279c925532d5cd8c)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>