bus: mhi: core: Check channel execution environment before issuing reset
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Bhaumik Bhatt <bbhatt@codeaurora.org>
commit 8e06e9fb990975a55b64c6e4c5b6b183962f849a

A client can attempt to unprepare certain channels for transfer even
after the execution environment they are supposed to run in has changed.
In the event that happens, the device need not be notified of the reset
and the host can proceed with clean up for the channel context and
memory allocated for it on the host as the device will no longer be able
to respond to such a request.

	Signed-off-by: Bhaumik Bhatt <bbhatt@codeaurora.org>
	Reviewed-by: Hemant Kumar <hemantk@codeaurora.org>
	Reviewed-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
Link: https://lore.kernel.org/r/1617311778-1254-8-git-send-email-bbhatt@codeaurora.org
	Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
(cherry picked from commit 8e06e9fb990975a55b64c6e4c5b6b183962f849a)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>