brcmsmac: Drop unnecessary NULL check after container_of
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Guenter Roeck <linux@roeck-us.net>
commit 34fe7038a3b33b4b50b1e948e005bf3db20b7a54

The parameter passed to ai_detach() is guaranteed to never be NULL
because it is checked by the caller. Consequently, the result of
container_of() on it is also never NULL, and a NULL check on it
is unnecessary. Even without that, the NULL check would still be
unnecessary because the subsequent kfree() can handle NULL arguments.
On top of all that, it is misleading to check the result of container_of()
against NULL because the position of the contained element could change,
which would make the check invalid. Remove it.

This change was made automatically with the following Coccinelle script.

@@
type t;
identifier v;
statement s;
@@

<+...
(
  t v = container_of(...);
|
  v = container_of(...);
)
  ...
  when != v
- if (\( !v \| v == NULL \) ) s
...+>

	Signed-off-by: Guenter Roeck <linux@roeck-us.net>
	Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
Link: https://lore.kernel.org/r/20210511235629.1686038-1-linux@roeck-us.net
(cherry picked from commit 34fe7038a3b33b4b50b1e948e005bf3db20b7a54)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>