iwlwifi: mvm: don't change band on bound PHY contexts
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Johannes Berg <johannes.berg@intel.com>
commit 8835a64f74c46baebfc946cd5a2c861b866ebcee

When we have a P2P Device active, we attempt to only change the
PHY context it uses when we get a new remain-on-channel, if the
P2P Device is the only user of the PHY context.

This is fine if we're switching within a band, but if we're
switching bands then the switch implies a removal and re-add
of the PHY context, which isn't permitted by the firmware while
it's bound to an interface.

Fix the code to skip the unbind/release/... cycle only if the
band doesn't change (or we have old devices that can switch the
band on the fly as well.)

	Signed-off-by: Johannes Berg <johannes.berg@intel.com>
	Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Link: https://lore.kernel.org/r/iwlwifi.20210612142637.e9ac313f70f3.I713b9d109957df7e7d9ed0861d5377ce3f8fccd3@changeid
	Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
(cherry picked from commit 8835a64f74c46baebfc946cd5a2c861b866ebcee)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>