iwlwifi: support loading the reduced power table from UEFI
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Luca Coelho <luciano.coelho@intel.com>
commit 9dad325f9d57508b154f0bebbc341a8528e5729c

This new feature allows OEMs to set a special reduced power table in a
UEFI variable, which we use to tell the firmware to change the TX
power tables.

Read the variable and store it in a dram block to pass it to the
firmware.  We do this as part of the PNVM loading flow.

	Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Link: https://lore.kernel.org/r/iwlwifi.20210621103449.259a33ba5074.I2e0bb142d2a9c412547cba89b62dd077b328fdc4@changeid
	Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
(cherry picked from commit 9dad325f9d57508b154f0bebbc341a8528e5729c)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>