mac80211: Fix monitor MTU limit so that A-MSDUs get through
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Johan Almbladh <johan.almbladh@anyfinetworks.com>
commit 79f5962baea74ce1cd4e5949598944bff854b166

The maximum MTU was set to 2304, which is the maximum MSDU size. While
this is valid for normal WLAN interfaces, it is too low for monitor
interfaces. A monitor interface may receive and inject MPDU frames, and
the maximum MPDU frame size is larger than 2304. The MPDU may also
contain an A-MSDU frame, in which case the size may be much larger than
the MTU limit. Since the maximum size of an A-MSDU depends on the PHY
mode of the transmitting STA, it is not possible to set an exact MTU
limit for a monitor interface. Now the maximum MTU for a monitor
interface is unrestricted.

	Signed-off-by: Johan Almbladh <johan.almbladh@anyfinetworks.com>
Link: https://lore.kernel.org/r/20210628123246.2070558-1-johan.almbladh@anyfinetworks.com
	Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 79f5962baea74ce1cd4e5949598944bff854b166)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>