iwlwifi: mvm: set replay counter on key install
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Johannes Berg <johannes.berg@intel.com>
commit a6dfbd040e26d0fa429545d37ca9d3dc8a80cf3b

When installing a (new) key, set the replay counter so that
after FW restart the firmware has the correct value of the
replay counters.

This doesn't have a large effect - for frames that reach
the driver, it will do a replay check, and when installing
a new key, the counter is normally zero to start with (not
for GTK though, if joining the BSS for the first time).

Since this only affects frames handled entirely by the FW,
and that's restricted to a few unicast management frames,
the only affect here is for those after a firmware restart.

	Signed-off-by: Johannes Berg <johannes.berg@intel.com>
	Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Link: https://lore.kernel.org/r/iwlwifi.20210802172232.1cedf2ca7bb6.I2e609c28eaa301436e6740f4f1beca838f69a96a@changeid
	Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
(cherry picked from commit a6dfbd040e26d0fa429545d37ca9d3dc8a80cf3b)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>