iwlwifi: acpi: support reading and storing WRDS revision 1 and 2
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Luca Coelho <luciano.coelho@intel.com>
commit 2a8084147bff7a1fe9f567ed39c340a6a3fc27ef

Change the SAR profile tables storage to revision 2 regardless of the
revision we read from ACPI.  Revision 2 is a superset of revision 1,
which is in turn a superset of revision 0, so they can all be stored
inside revision 2.

Add support for reading and storing also revisions 1 and 2, whose only
difference is the number of chains and number of sub-bands.  So most
of the code revolves around passing different chain and sub-band
sizes.

With this patch we still pass only revision 0 to the firmware, but
that will be changed in a separate patch.

	Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Link: https://lore.kernel.org/r/iwlwifi.20210805141826.2edad09415c8.I36297aef90a9ec6a3ea1bf1a151a62f272826d59@changeid
	Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
(cherry picked from commit 2a8084147bff7a1fe9f567ed39c340a6a3fc27ef)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>