crypto: qat - make PFVF send and receive direction agnostic
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Marco Chiappero <marco.chiappero@intel.com>
commit 028042856802c4731c6afebe15d95fed4d39a614

Currently PF and VF share the same send and receive logic for the PFVF
protocol. However, the inner behaviour still depends on the specific
direction, requiring a test to determine the if the sender is a PF or a
VF. Moreover the vf_nr parameter is only required for PF2VF messages and
ignored for the opposite direction.

Make the GEN2 send and recv completely direction agnostic, by calculating
and determining any direction specific input in the caller instead, and
feeding the send and the receive functions with the same arguments for
both PF and VF. In order to accommodate for this change, the API of the
pfvf_ops send and recv has been modified to remove any reference to vf_nr.

	Signed-off-by: Marco Chiappero <marco.chiappero@intel.com>
	Reviewed-by: Giovanni Cabiddu <giovanni.cabiddu@intel.com>
	Reviewed-by: Fiona Trahe <fiona.trahe@intel.com>
	Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
(cherry picked from commit 028042856802c4731c6afebe15d95fed4d39a614)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>