USB: host: uhci: remove dentry pointer for debugfs
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Greg Kroah-Hartman <gregkh@linuxfoundation.org>
commit 5649d86f537887c2be88689986ec4dd493d4babe

There is no need to keep the dentry pointer around for the created
debugfs file, as it is only needed when removing it from the system.
When it is to be removed, ask debugfs itself for the pointer, to save on
storage and make things a bit simpler.

And, no one noticed that a __maybe_unused dentry * in uhci_start()
really was unused, so remove that as it's obviously not needed, and
hasn't been for quite some time.

	Cc: Alan Stern <stern@rowland.harvard.edu>
	Cc: linux-usb@vger.kernel.org
Link: https://lore.kernel.org/r/20210216144645.3813043-4-gregkh@linuxfoundation.org
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 5649d86f537887c2be88689986ec4dd493d4babe)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>