thunderbolt: Add support for maxhopid XDomain property
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Mika Westerberg <mika.westerberg@linux.intel.com>
commit 46b494f286812a88caba28dd0810cf3a55747431

USB4 inter-domain spec mandates that the compatible hosts expose a new
property "maxhopid" that tells the connection manager on the other side
what is the maximum supported input HopID over the connection. Since
this is depend on the lane adapter the cable is connected it needs to be
filled in dynamically.

For this reason we take a copy of the global properties and fill then
for each XDomain connection upon first connect, and then keep updating
it if the generation changes as services are being added/removed. We
also take advantage of this copy to fill in the hostname.

We also expose this maxhopid as an attribute under each XDomain device.

While there drop kernel-doc entry for property_lock which seems to be
left there when the structure was originally introduced.

	Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 46b494f286812a88caba28dd0810cf3a55747431)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>