USB: serial: iuu_phoenix: remove redundant variable 'error'
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Colin Ian King <colin.king@canonical.com>
commit 315e2811f58b807e6e76b7385e3d7b970f4562d2

The variable error is initialized to 0 and is set to 1 this
value is never read as it is on an immediate return path. The
only read of error is to check it is 0 and this check is always
true at that point of the code. The variable is redundant and
can be removed.

Addresses-Coverity: ("Unused value")
	Signed-off-by: Colin Ian King <colin.king@canonical.com>
	Signed-off-by: Johan Hovold <johan@kernel.org>
(cherry picked from commit 315e2811f58b807e6e76b7385e3d7b970f4562d2)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>