USB: serial: opticon: fix TIOCGSERIAL implementation
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Johan Hovold <johan@kernel.org>
commit aa6a45850224fc19dbc6058598a9dc57db45b530

TIOCSSERIAL is a horrid, underspecified, legacy interface which for most
serial devices is only useful for setting the close_delay and
closing_wait parameters.

The xmit_fifo_size parameter could be used to set the hardware transmit
fifo size of a legacy UART when it could not be detected, but the
interface is limited to eight bits and should be left unset when not
used.

Similarly, baud_base could be used to set the UART base clock when it
could not be detected but might as well be left unset when it is not
known.

The close_delay and closing_wait parameters returned by TIOCGSERIAL are
specified in centiseconds (not jiffies). The driver does not yet support
changing these, but let's report back the default values actually used
(0.5 and 30 seconds, respectively).

Fixes: faac64ad9c7b ("USB: serial: opticon: add serial line ioctls")
	Reviewed-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
	Signed-off-by: Johan Hovold <johan@kernel.org>
(cherry picked from commit aa6a45850224fc19dbc6058598a9dc57db45b530)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>