usb: dwc2: Fix gadget DMA unmap direction
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Phil Elwell <phil@raspberrypi.com>
commit 75a41ce46bae6cbe7d3bb2584eb844291d642874

The dwc2 gadget support maps and unmaps DMA buffers as necessary. When
mapping and unmapping it uses the direction of the endpoint to select
the direction of the DMA transfer, but this fails for Control OUT
transfers because the unmap occurs after the endpoint direction has
been reversed for the status phase.

A possible solution would be to unmap the buffer before the direction
is changed, but a safer, less invasive fix is to remember the buffer
direction independently of the endpoint direction.

Fixes: fe0b94abcdf6 ("usb: dwc2: gadget: manage ep0 state in software")
	Acked-by: Minas Harutyunyan <Minas.Harutyunyan@synopsys.com>
	Cc: stable <stable@vger.kernel.org>
	Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Link: https://lore.kernel.org/r/20210506112200.2893922-1-phil@raspberrypi.com
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 75a41ce46bae6cbe7d3bb2584eb844291d642874)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>