usb: dwc3: omap: improve extcon initialization
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Marcel Hamer <marcel@solidxs.se>
commit e17b02d4970913233d543c79c9c66e72cac05bdd

When extcon is used in combination with dwc3, it is assumed that the dwc3
registers are untouched and as such are only configured if VBUS is valid
or ID is tied to ground.

In case VBUS is not valid or ID is floating, the registers are not
configured as such during driver initialization, causing a wrong
default state during boot.

If the registers are not in a default state, because they are for
instance touched by a boot loader, this can cause for a kernel error.

	Signed-off-by: Marcel Hamer <marcel@solidxs.se>
Link: https://lore.kernel.org/r/20210427122118.1948340-1-marcel@solidxs.se
	Cc: stable <stable@vger.kernel.org>
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit e17b02d4970913233d543c79c9c66e72cac05bdd)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>