usb: typec: mux: Fix matching with typec_altmode_desc
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Bjorn Andersson <bjorn.andersson@linaro.org>
commit acf5631c239dfc53489f739c4ad47f490c5181ff

In typec_mux_match() "nval" is assigned the number of elements in the
"svid" fwnode property, then the variable is used to store the success
of the read and finally attempts to loop between 0 and "success" - i.e.
not at all - and the code returns indicating that no match was found.

Fix this by using a separate variable to track the success of the read,
to allow the loop to get a change to find a match.

Fixes: 96a6d031ca99 ("usb: typec: mux: Find the muxes by also matching against the device node")
	Reviewed-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
	Cc: stable <stable@vger.kernel.org>
	Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Link: https://lore.kernel.org/r/20210516034730.621461-1-bjorn.andersson@linaro.org
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit acf5631c239dfc53489f739c4ad47f490c5181ff)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>