Revert "usb: typec: mux: Remove requirement for the "orientation-switch" device property"
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Heikki Krogerus <heikki.krogerus@linux.intel.com>
commit 7f7d0afe1d479990b712ecb494257dcd706a8016

This reverts commit acad3e9c7250c5fd20d9778a163f2adc95de38f5.

The device property that can be used to identify the device
class/type of the remote port parent when device graph is
used is always needed after all. Without it there is no real
way to know is the requested connection actually described
in the device graph or not.

If the connection is described in the device graph but the
device instance is still missing for what ever reason, the
code defers probe for now.

Adding a comment to the code to explain this.

	Reviewed-by: Li Jun <jun.li@nxp.com>
	Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Link: https://lore.kernel.org/r/20210602112253.70200-1-heikki.krogerus@linux.intel.com
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 7f7d0afe1d479990b712ecb494257dcd706a8016)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>