usb: dwc3: gadget: Fix dwc3_calc_trbs_left()
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Thinh Nguyen <Thinh.Nguyen@synopsys.com>
commit 51f1954ad853d01ba4dc2b35dee14d8490ee05a1

We can't depend on the TRB's HWO bit to determine if the TRB ring is
"full". A TRB is only available when the driver had processed it, not
when the controller consumed and relinquished the TRB's ownership to the
driver. Otherwise, the driver may overwrite unprocessed TRBs. This can
happen when many transfer events accumulate and the system is slow to
process them and/or when there are too many small requests.

If a request is in the started_list, that means there is one or more
unprocessed TRBs remained. Check this instead of the TRB's HWO bit
whether the TRB ring is full.

Fixes: c4233573f6ee ("usb: dwc3: gadget: prepare TRBs on update transfers too")
	Cc: <stable@vger.kernel.org>
	Acked-by: Felipe Balbi <balbi@kernel.org>
	Signed-off-by: Thinh Nguyen <Thinh.Nguyen@synopsys.com>
Link: https://lore.kernel.org/r/e91e975affb0d0d02770686afc3a5b9eb84409f6.1629335416.git.Thinh.Nguyen@synopsys.com
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 51f1954ad853d01ba4dc2b35dee14d8490ee05a1)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>