HID: i2c-hid: Prefer asynchronous probe
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Douglas Anderson <dianders@chromium.org>
commit eafb2203626aedb599cc80aadcba533c917b53b0

Adding printouts to the i2c_hid_probe() function shows that it takes
quite some time.  It used to take about 70 ms, but after commit
eef4016243e9 ("HID: i2c-hid: Always sleep 60ms after I2C_HID_PWR_ON
commands") it takes about 190 ms.  This is not tons of time but it's
not trivial.  Because we haven't yet specified that we'd prefer
asynchronous probe for this driver then, if the driver is builtin to
the kernel, we'll wait for this driver to finish before we start
probes for more drivers.  Let's set the flag to enable asynchronous
for this driver so that other drivers aren't blocked from probing
until we finish.

Since this driver can be configured as a module and modules are
always asynchronously probed this is quite a safe change and will
benefit anyone who has a reason to build this driver into the kernel
instead of using it as a module.

[jkosina@suse.cz: drop spurious whitespace addition]
	Signed-off-by: Douglas Anderson <dianders@chromium.org>
	Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit eafb2203626aedb599cc80aadcba533c917b53b0)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>