dm space maps: improve performance with inc/dec on ranges of blocks
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Joe Thornber <ejt@redhat.com>
commit be500ed721a6ec8d49bf0814c277ce7162acee0e

When we break sharing on btree nodes we typically need to increment
the reference counts to every value held in the node.  This can
cause a lot of repeated calls to the space maps.  Fix this by changing
the interface to the space map inc/dec methods to take ranges of
adjacent blocks to be operated on.

For installations that are using a lot of snapshots this will reduce
cpu overhead of fundamental operations such as provisioning a new block,
or deleting a snapshot, by as much as 10 times.

	Signed-off-by: Joe Thornber <ejt@redhat.com>
	Signed-off-by: Mike Snitzer <snitzer@redhat.com>
(cherry picked from commit be500ed721a6ec8d49bf0814c277ce7162acee0e)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>