irqchip/gic-v3: Spell out when pseudo-NMIs are enabled
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-365.el8
commit-author Alexandru Elisei <alexandru.elisei@arm.com>
commit 4e594ad1068ea1db359d6161f580f03edecf6cb0

When NMIs cannot be enabled, the driver prints a message stating that
unambiguously. When they are enabled, the only feedback we get is a message
regarding the use of synchronization for ICC_PMR_EL1 writes, which is not
as useful for a user who is not intimately familiar with how NMIs are
implemented.

Let's make it obvious that pseudo-NMIs are enabled. Keep the message about
using a barrier for ICC_PMR_EL1 writes, because it has a non-negligible
impact on performance.

	Signed-off-by: Alexandru Elisei <alexandru.elisei@arm.com>
	Signed-off-by: Marc Zyngier <maz@kernel.org>
Link: https://lore.kernel.org/r/20200912153707.667731-2-alexandru.elisei@arm.com
(cherry picked from commit 4e594ad1068ea1db359d6161f580f03edecf6cb0)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>